feat: update BaseSampler interface
main changes:
- introduce 'sample_batch' abstract method, which allows the client code to
  sample a batch of points in the parameter space;
- remove 'single_sample' abstract method.

The main reason about this change is non-functional. Many sampling strategies
can optimize the computation if they are not constrained to sample a single
point at a time.

The new implementation of the samplers exploit such gained flexibility of the
interface, and exploit vectorized computation to speed up the performances.

The main change happened to the Halton sampler and the RSequence sampler. Some
detail on the improved performances can be found, respectively, at the following
links:

- https://github.com/bancaditalia/black-it/pull/136#issuecomment-1082357257
- https://github.com/bancaditalia/black-it/pull/136#issuecomment-1094041283

The rest of the changes just updates the tests and the relevant documentation.