Change random seed interface
The random seed structure of the calibrator and of the samplers has been changed.
This change was already present in #2 but I am reproducing it here to ease code review.

Now the seed given to the calibrator is used to initialise the seeds of the samplers in a random (yet reproducible) way,
and it is also used to initialise a random generator hat gives different seeds to the model simulations.

This commit also fixes GPy non-deterministic behaviour by setting global seeds in numpy and python