ci: install package and relevant development dependencies for pylint step
Pylint requires the package to be installed in the current Python environment,
as well as any other development dependencies used in tests and examples.

For example, 'mesa' is used in the examples, and was missing from the list.

Therefore, we add a 'pip install .' command in the step for the preparation
of the environment for the Pylint checks, and then the dependency 'mesa'
in the following command.