fix: make random_state Optional[int]
the rationale behind this change is that it is common in the APIs of scientific
libraries to have the default value of the seed parameter set to None, which means
the random seed is set according to a source of entropy depending on the
operating system (e.g. on UNIX this is the pseudodevice /dev/urandom).
Moreover, the use of '0' as default value would prevent the possibility to
'not set' the seed, which sometimes is desirable.