feat: add 'global_minimum_across_samplers' option to ParticleSwarmSampler
This commit adds a new boolean option to the constructor of ParticleSwarmSampler,
namely global_minimum_across_samplers. This option allows to switch between
two types of behaviours of the particle swarm sampler:

If `global_minimum_across_samplers` is False, then the velocity update rule, for the global
minimum part, uses the current position of the particle that in its history
has sampled, so far, the best set of parameters.

Else, if `global_minimum_across_samplers` is True, then the velocity update rule
returns the point in parameter space that achieved the minimum loss.
Note that this point could have been sampled by a different sampler than "self".