fix: XGBoostSampler._clip_losses was clipping when not needed
This commit fixes a subtle bug in XGBoostSampler._clip_losses. The check
on large/small floats makes use of np.where, which returns a tuple where
the conditions provided in input hold. In case of one condition, it returns
a tuple of one element. However, right below, the result was used as if it
were the actual list of items that satisfied the condition, rather than
as it is wrapped with a tuple. This makes the check always to fail, and to
clip even when not needed.

The proposed change correctly unpacks the tuple returned by np.where.