test: fix GslDivLoss.discretize doctest on Windows
This change is motivated by the fact that, for some reason, the representation
of Numpy arrays in Windows also include their dtype.

For example, running the following code snippet on Linux:

    GslDivLoss.discretize(
        [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
        nb_values = 3,
        start_index = 1,
        stop_index = 10
    )

Gives:

    array([1, 1, 1, 2, 2, 2, 2, 3, 3, 3])

While on Windows, the same code produces the following representation:

    array([1, 1, 1, 2, 2, 2, 2, 3, 3, 3], dtype=int64)

This of course led to the failure of the doctest, since it is based on the representation of the object being tested for equality.
Using .tolist() was a relatively straighforward (although less elegant) solution.