fix: tox mypy job testenv installs the package
this is needed because some dependencies already have type hinting that otherwise
would not be exploited; and instead of listing them as dependencies, we use
a more flexible configuration that does not require maintainance in case of dependency update.

It will increase the running time of the mypy tox task a bit though.