lint: fix flake8-executable errors
To reproduce the errors:

```
ruff check --select "EXE" black_it tests examples scripts
```

The (unique) error was fixed manually.

The black_it/calibrator.py module contained a shebang, but it was not executable. It used to be executable, so probably when the executable part was removed (i.e. __name__ == "__main__"), the shebang was not updated accordingly.