lint: fix flake8-no-pep420 errors
To reproduce the errors:

```
ruff check --fix --select "INP" black_it tests examples scripts
```

The error were fixed by adding __init__.py whenever was needed.
In two cases, namely "examples/docker-sir.py" and "examples/models/simple_models.py", we ignored the INP001 errors.

In particular 'docker-sir.py' and 'main.py' are supposed to be run as script, so it does not make sense to have them as part of a package.

For some reason, simple_models.py was not considered part of the 'models' package, even after adding __init__.py; hence we added it to the ignore list.