fix: update coordinate_filters argument type annotation in BaseLoss.__init__
A good practice is to use the coarsest type as argument function, so to make it usable in more ways.

In particular, in this commit, the argument coordinate_filters in the base constructor of the class BaseLoss has been changed from 'list' to 'Sequence'. This allows to use coordinate_filters as tuples. It is also a more correct type hint, since coordinate_filters should not change during the lifetime of BaseLoss.