fix!: replace GPy with scikit-learn in GaussianProcessSampler
BREAKING CHANGE: This commit updates the implementation of GaussianProcessSampler. Instead of relying on GPy, now we use the implementation of Gaussian Processes provided by the scikit-learn package.

The change introduces a different behaviour of the GaussianProcessSampler, even if the same random seed is provided. This is witnessed by the updated expected sampled parameters and losses in tests/test_calibrator.py.

Another notable difference is in the keyword arguments for the GaussianProcessSampler class constructor. In particular:
- the keyword argument 'max_iters' has been removed
- the keyword argument 'jitter' has been added.