build!: remove 'ipykernel' dependency
The dependency `ipykernel` is not used directly from any module in `black-it`,
hence it should not be declared as main dependency of the package. This commit
removes it from the section `tool.poetry.dependencies` in `pyproject.toml`, and
recomputes `poetry.lock`.

Potentially breaking change if packages dependent of `black-it` use `ipykernel`
directly (unlikely).