build: remove ipykernel
The dependency 'ipykernel' is not a direct dependency of the package.

This can be verified by grepping import statements involving ipykernel package:

    $ grep -r "import ipykernel"
    $ grep -r "from ipykernel import"

Both commands do not return any result.
Note: this is not a complete approach to detect direct uses of a package, but it
      is a *reasonably* complete approach.

This commit removes the line for ipykernel in pyproject.toml. Then, it runs
`poetry lock`. Note that no dependency has changed in poetry.lock, meaning that
ipykernel is a subdependency of another direct dependency (e.g. see ipywidgets
dependencies).