env: introduce support for python 3.12, but drop support for python 3.8 due to scipy
As of 2024-02, Python 3.8 is still going to be supproted for 6 months (until
2024-10, see https://devguide.python.org/versions).

At the same time, Python 3.12 was released four months ago, on 2023-10, and will
be supported for five years.

We cannot support both, because scipy 1.10 does not work with python 3.12, and
scipy 1,11 drops python 3.8.

Having to make a choiche, it is probably better to go with the more recent
versions, thus 3.8 has to go down the drain.