deps: update numpy and statsmodels in preparation to solving python3.12 incompatibilities
The issue with numpy is documented here, and officially solved with numpy 1.26:
https://numpy.org/devdocs/reference/distutils_status_migration.html

The upgrade on statsmodel probaby uncovers a bug in its MacOS build: the values
computed by MethodOfMomentsLoss are slightly different there, and lead to a
different result in the test.

Let's accept it for now; we can eventually file a bug on statsmodel later.