tests: fix PytestRemovedIn8Warning ("support for nose tests is deprecated and will be removed in a future release")
Before this change, the test log contained 37 entries like this:
    PytestRemovedIn8Warning: Support for nose tests is deprecated and will be removed in a future release.

After this change, on my machine the test log shrinks from 1672 to 1405 lines
(16% less).

At some point in time, we will migrate to Pytest 8.x, and this would be an
hindrance.

A reference for the fix can be found at:
    https://docs.pytest.org/en/stable/deprecations.html#support-for-tests-written-for-nose

Affected tests were found with:
    rg "def (setup|teardown)\(" tests