scripts: show that setup.cfg and .ruff.toml currently target different python versions
Currently mypy targets python 3.8 (in setup.cfg) and ruff targets 3.9 (in
.ruff.toml).

This commit introduces the necessary logic to show the mismatch. The next commit
will fix this error and ensure that form now on the versions change in lockstep.

The dependency on tomli in pyproject.toml can be removed once the minimum
supported version of the project becomes 3.11.