tox: use commands from the Makefile where possible, without duplicating them
This commit starts from the observation that the commands in tox.ini are often
a duplicate of the ones contained in the Makefile. This reduces the confidence a
developer can have in his local tests, because there is no automatic process
that guarantees that the CI will run the exact same tests he has run locally.

This change goes through tox.ini and compares the commands with the ones in the
Makefile. When they are completely equal, the command in tox.ini are replaced
with Makefile invocations.

This reduces the maintenance work in the future, because a change will not need
to be kept in sync between two places. Moreover, it gives developers more
confidence when changing the code base, because their local tests and the ones
in the CI will be better aligned.