docs: the minimum supported python version is currently 3.9 and not 3.8
We really should add an automatic check here, but it is not straightforward.

On one hand, it is is easy to get a specifier for the supported python versions
in a standards-compliant way:

    >>> import importlib.metadata as implib_meta
    >>> from packaging.specifiers import SpecifierSet
    >>> SpecifierSet(implib_meta.metadata("black_it")["Requires-Python"])
    <SpecifierSet('<3.13,>=3.9')>

On the other hand, a SpecifierSet only offers membership tests: there is no
concept of ordering and no support for getting the minimum allowed version.